pyproject.toml
[project]
name = "deeplearn24"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "typing-extensions>=4.12.2",
    "httpx",
    "levenshtein>=0.25.1",
]
requires-python = "<3.12,>=3.11"
readme = "README.md"
license = {text = "GPLv3"}

[project.optional-dependencies]
handout_notebook = [
    "torchvision==0.2.1",
    "pillow==8.3.1",
]
doc_ufcn = [
    "opencv-python>=4.8.1.78",
    "numpy>=1.24.1",
    "doc-ufcn>=0.1.8",
    "scikit-image>=0.24.0",
    "pillow>=10.4.0",
    "deskew>=1.5.1",
    "tqdm>=4.66.4",
    "jupyter>=1.0.0",
    "python-docx>=1.1.2",
    "mlflow>=2.14.2",
    "pytesseract>=0.3.10",
    "jiwer>=3.0.4",
    "plotly>=5.22.0",
]
handout_model = [
    "tensorflow==2.15"
]
trocr = [
    "datasets>=2.20.0",
    "mlflow>=2.14.2",
    "transformers[torch]>=4.42.4",
    "torch>=2.3.1",
    "jiwer>=3.0.4",
    "evaluate>=0.4.2",
    "setuptools>=71.1.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.5.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["I", "TCH"]